Update .sln and Resources.resx files
Updated the solution file to include a new `ExtensibilityGlobals` section with a `RESX_SortFileContentOnSave` property set to `True` and a defined `SolutionGuid`. Modified the `Resources.resx` file to correct the XML declaration and to add a new resource string `ErrorThereWasAnErrorWithJsonSerialization`. Additionally, reorganized the placement of the newly added resource string for better readability.