Refactor EnumHelper.cs and update metadata
Updated the header comment to reflect the new last modified date. Added a new `using` directive for `DotNetTips.Spargine.Core.Devices`. Modified the `Information` attribute on the `GetValues` method, changing `BenchMarkStatus` from `Completed` to `CheckPerformance`. Introduced a new variable `itemCount` to store the length of `allValues` and updated the `for` loop to use `itemCount` instead of directly accessing `allValues.Length`.