Mark methods in CollectionBenchmark as obsolete
The `CollectionBenchmark` class in `CollectionBenchmark.Person.cs`
and `CollectionBenchmark.PersonRecord.cs` has been updated to mark
three methods as obsolete. These methods are:
1. `GetPersonRefCollection()`
2. `GetPersonValCollection()`
3. `GetPersonRecordCollection()`

Each method now has an `[Obsolete]` attribute with a message
indicating that they should be replaced with their respective
array-returning methods (`GetPersonRefArray()`, `GetPersonValArray()`,
and `GetPersonRecordArray()`) and that they will be removed at the
end of 2024.