Update metadata attributes and modify last modified dates
Updated `JsonSerialization.cs` to include a new `[Information]`
attribute with a documentation URL for the `JsonSerialization`
class. Updated the "Last Modified On" date to "10-26-2024".

In `DictionaryExtensions.cs`, updated the "Last Modified On"
date to "10-26-2024" and added a new `[Information]` attribute
with a documentation URL for the `DictionaryExtensions` class.
Removed the `Documentation` field from the `[Information]`
attributes for the following methods:
- `HasItems`
- `ToConcurrentDictionary`
- `ToDelimitedString`
- `ToFrozenDictionary`
- `ToImmutableSortedDictionary`
- `ToReadOnlyCollection`
- `ToReadOnlyDictionary`
- `ToSortedDictionary`
- `Upsert`