Refactor ListExtensions methods and update metadata
Updated the metadata comment in `ListExtensions.cs` to reflect a new modification date (10-29-2024). Modified the `Information` attribute for `AddFirst` and `AddLast` methods, changing `BenchMarkStatus` from `NotRequired` to `CheckPerformance`. Refactored `AddFirst` to validate the collection and insert the item at the beginning. Refactored `AddLast` to validate the collection and add the item to the end.