Optimize parameter passing with 'in' keyword
Updated `ToDelimitedString` in `DictionaryExtensions` to use `in` for `delimiter` parameter.
Updated `RunProcessAndIgnoreOutput` and `RunProcessAndReadOutput` in `ProcessExtensions` to use `in` for `timeout` parameter.
Updated `Last Modified On` date in `ProcessExtensions.cs` to `11-13-2024`.

Optimize parameter passing with 'in' keyword

Updated methods in DictionaryExtensions, ProcessExtensions, and StringBuilderExtensions to use the 'in' keyword for performance improvement by passing parameters by reference. Also updated the 'Last Modified On' date in ProcessExtensions.cs.