Update SHA256PasswordHasher comments and attributes
Updated the file header comment in `SHA256PasswordHasher.cs` to reflect a new modification date (11-26-2024). Changed the `Status` in the `Information` attribute for the `SHA256PasswordHasher` class from `NeedsDocumentation` to `Available` and added a new `Documentation` URL. Removed the `Documentation` URL from the `Information` attributes for both the `HashPassword` and `VerifyHashedPassword` methods.