Update DotNetTips.Spargine to version 2025.8.1.6
Updated the version number of the DotNetTips.Spargine library for Microsoft .NET 8 from `2025.8.1.5` to `2025.8.1.6` across multiple project files and the README.md file. This includes updating the version number in the title of the `README.md` file and the `AssemblyVersion`, `FileVersion`, and `Version` properties in the `.csproj` files for various components.