Add TODOs for future return type changes to void
Updated last modified dates to January 17, 2025, in `ObservableList.cs` and `CollectionExtensions.cs`. Added TODO comments in `ObservableList.cs` and `CollectionExtensions.cs` to change the return type of `Add` and `AddRange` methods to `void` at the end of 2025. Added a TODO comment in `CollectionExtensions.cs` to add a new method `AddRangeIfNotExists`. Added a TODO comment in `ListExtensions.cs` to change the return type of `AddRangeIfNotExists` to `void` at the end of 2025.