Update file header date and modify Information attributes
Updated the `Last Modified On` date in the file header comment from `10-09-2024` to `02-08-2025`. Changed the `Information` attribute for the `PathHelper` class to include a `Documentation` URL, updating the status from `Status.NeedsDocumentation` to `Status.Available` with the URL `https://bit.ly/SparginePathHelper`. Removed the `Documentation` URL from the `Information` attributes of the `CombinePaths`, `EnsureTrailingSlash`, `HasInvalidFilterChars`, `PathContainsWildcard`, and `PathHasInvalidChars` methods.